merge-base: fp experiment
Update the condition to stop traversal to also pay attention to the
first-parent chain.  I am not sure if this is a more complex and
unnecessary way to implement the same as only traversing the first
parent chain to find merge bases.

And by definition, if we find merge bases only by traversing the
first parent chain, that would defeat the "criss-cross" mitigation,
which is not a good idea unless the history's shape is very well
controlled (e.g. by employing the topic-branch workflow correctly).