freshen_object(): factor out a helper function
Callers often call has_sha1_file() to see if an object already
exists in the object store to short-cut writing the same object out.

It is safe to do so as long as such an object will be made reachable
from an object that is newly created by these callers, as the youth
propagates from the reachability chain for the purpose of protecting
young objects that are unreachable from refs and other anchoring
points from getting pruned.  And any object that is "newly created"
goes through write_sha1_file() API and is marked as "young" by
freshen_packed_object() and freshen_loose_object() helper functions.

There however are some codepaths that bypass write_sha1_file() API,
and stops after seeing that the object already exists in the object
store.  They need to have a way to freshen the object they would
have written if it weren't there already.

Refactor the idiomatic calls to freshen_{packed,loose}_object() into
a single helper function freshen_object() and export it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>