commit: remove redundant check for active_cache_changed
It is idiomatic to say "if the cache-tree is not fully valid,
then call update_cache_tree() to update it", but this old codepath
had one extra condition to call update_cache_tree(), which was "if
anything in the index changed, call update_cache_tree()".

This is unnecessary.  Whenever any change is made to an index entry,
the cache-tree entries that cover the path are invalidated and the
cache-tree becomes "not fully valid" at that point.  Remove it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>