cache-tree: retire cache_tree_fully_valid() API function
All callers of this helper use it like so:

    if (!cache_tree_fully_valid(...))
	cache_tree_update(...);

to optimize out the call to cache_tree_update() when the cache-tree
structure is already known to be valid.  However, cache_tree_update()
function itself already has an equivalent check that short-cuts the
recomputation of the tree object name for each subdirectory.

Signed-off-by: Junio C Hamano <gitster@pobox.com>