cache-tree: freshen the tree object at the top level
"git write-tree" that is asked to write out a tree object out of the
index may not actually write the tree object anew, when there
already is the same tree object in the object store.  The actual
writing of the tree object is done as a side effect of updating the
cache tree entries fully in the index, and the optimization to skip
writing the tree out is done via "has_sha1_file()", i.e. "does the
tree already exist in the object store?"

After asking "git write-tree" to write a tree out, the caller may
not make it reachable from any ref or other anchoring point, which
makes the tree object subject to pruning.  Call freshen_object() on
the top-level tree object to make sure we mark the tree "young" to
protect it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>