apply: file commited with CRLF should roundtrip diff and apply
When a file had been commited with CRLF and core.autocrlf is true,
the following does not roundtrip, `git apply` fails:

    printf "Added line\r\n" >>file &&
    git diff >patch &&
    git checkout -- . &&
    git apply patch

Before applying the patch, the file from working tree is converted
into the index format (clean filter, CRLF conversion, ...)  Here,
when commited with CRLF, the line endings should not be converted.

Analyze the patch if there is any context line with CRLF, or if any
line with CRLF is to be removed.

If yes, the new flag has_crlf is set in "struct patch", and two
things will happen:

    - read_old_data() will not convert CRLF into LF by calling
      convert_to_git(..., SAFE_CRLF_KEEP_CRLF);
    - The WS_CR_AT_EOL bit is set in the "white space rule",
      CRLF are no longer treated as white space.

Reported-by: Anthony Sottile <asottile@umich.edu>
Signed-off-by: Torsten Bögershausen <tboegi@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>