usage_with_options: omit double new line on empty option list
Currently the worktree command gives its usage, when no subcommand is
given. However there are no general options, all options are related to
the subcommands itself, such that:

|    $ git worktree
|    usage: git worktree add [<options>] <path> [<branch>]
|       or: git worktree list [<options>]
|       or: git worktree lock [<options>] <path>
|       or: git worktree prune [<options>]
|       or: git worktree unlock <path>
|
|
|    $

Note the two empty lines at the end of the usage string. This is because
the toplevel usage is printed with an empty options list.

Only print one new line after the usage string if the option list is empty.

Signed-off-by: Stefan Beller <sbeller@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>