show: add --follow-symlinks option for <rev>:<path>
Add a --follow-symlinks option that'll resolve symlinks to their
targets when the target is of the form <rev>:<path>.

Without it, git will show the path of the link itself if the symlink
is the leaf node of <path>, or otherwise an error if some component
of <path> is a symlink to another location in the repository. With
the new --follow-symlinks option both will be resolved to their
target, and its content shown instead.

Signed-off-by: Michael Vogt <mvo@ubuntu.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>